feat: add follow button to user profile card components
