+++ projects/commudle-admin/src/app/feature-modules/community-channels/components/discussion-community-channel/community-channel-message/community-channel-message.component.ts
@@ -6,7 +6,6 @@ import {
   OnDestroy,
   OnInit,
   Output,
-  SimpleChanges,
   TemplateRef,
   ViewChild,
 } from '@angular/core';
