+++ projects/shared-components/profile-cards/user-profile-card-large/user-profile-card-large.component.html
@@ -1,15 +1,18 @@
 <a [routerLink]="['/users', user.username]">
   <div class="user-profile-large d-flex align-items-center">
     <div class="avatar">
-      <img alt="{{user.name}}" src="{{user.avatar}}" title="{{user.name}}">
+      <img alt="{{user.name}}" src="{{user.avatar}}" title="{{user.name}}" />
     </div>
     <div class="details">
       <div class="name d-flex align-items-center">
         {{user.name | truncatetext:maxNameLength}}
-        <img *ngIf="user.is_expert" alt="Expert Tick" src="/assets/images/Expert-Tick.svg" title="Expert Tick">
+        <img *ngIf="user.is_expert" alt="Expert Tick" src="/assets/images/Expert-Tick.svg" title="Expert Tick" />
       </div>
       <p class="username">@{{user.username | truncatetext:maxUserNameLength}}</p>
       <p class="designation">{{(user.designation || '--') | truncatetext:maxDesignationLength}}</p>
     </div>
   </div>
 </a>
+<div class="followers">
+  <app-user-follow [username]="user.username" class="follow-btn"></app-user-follow>
+</div>
