+++ projects/shared-components/profile-cards/user-profile-card-large/user-profile-card-large.component.scss
@@ -1,8 +1,11 @@
 @import '/projects/commudle-admin/src/assets/styles/variables';
 
-a, a:hover, a:focus, a:active {
-  text-decoration: none;
+a,
+a:hover,
+a:focus,
+a:active {
   color: inherit;
+  text-decoration: none;
 }
 
 p {
@@ -13,10 +16,10 @@ p {
   background-color: $white;
 
   .avatar {
-    max-width: 56px;
-    max-height: 56px;
-    min-width: 56px;
     border-radius: 50%;
+    max-height: 56px;
+    max-width: 56px;
+    min-width: 56px;
     overflow: hidden;
 
     img {
@@ -28,28 +31,33 @@ p {
     margin-left: $space-16;
 
     .name {
+      color: $black;
       font-size: $fs-18;
       line-height: 21px;
-      color: $black;
 
       img {
+        height: 16px;
         margin-left: $space-4;
         width: 16px;
-        height: 16px;
       }
     }
 
     .username {
-      font-size: $fs-h6;
-      line-height: 14px;
       color: rgba($gunmetal, 0.85);
+      font-size: $fs-h6;
+      line-height: 12px;
+      margin-top: $space-4;
     }
 
     .designation {
-      margin-top: $space-4;
+      color: $black;
       font-size: $fs-h5;
       line-height: 16px;
-      color: $black;
+      margin-top: $space-4;
     }
   }
+}
+
+.followers {
+  margin: $space-8 0 0 $space-72;
 }
